Fixed the path to the snapwebsites.org, it's HTTPS now.
Enhanced the --save-cell description in snapdb.
Use the pointer_t for advgetopt in snapdb.h instead of redefining it.
Added a parameter to the form error log for easier debugging.
Removed the WARNING of the server manager about typing the wrong name since Doug added a drop down and thus you can't make such a mistake.
Fixed a few #include for advgetopt which should always have used the path.