Enhanced glob_dir to support std::string and char const *.
Made use of the raii_pointer_deleter to track the glob_t object.
Updated the manager.cpp and journal.cpp to the new version of glob_dir.
Documented glob_dir and journal_list.
Changed when the journal_list becomes "finished": which is now when the done() function is called (because once marked done we should never reuse that pointer!)