Fixed the JS to remove old feedback messages, the while() was not working.
Fixed the tabs in the snapmanager, they look decent now and match their surrounding which makes it easier to see which is selected.
Made small adjustments to the code to get the correct classes to help with the CSS of the tabs and the tab panels.
Added a snapmanager.cgi widget (a link really) to view watchdog results (although the link destination doesn't work yet...)