Made code -Weffc++ compatible.
Removed the cell post ++ and -- because it returns itself and effective C++ says we should not have such. We're not using that feature anyway.
Converted a struct + function in a lambda now that we have C++11.
Check for /var/mail/root emails, if present show a warning in the "test_mta" field (snapwatchdog plugin of snapmanager.cgi).