Fixed the f_file and f_func whether we use the syslog() is used or not.
Test the string pointer for NULL in case we log char * or wchar_t *.
Log the URI when snap.cgi gets an I/O error to see when it happens.
Added a script to show how to unblock one IP in the firewall.
Created a document showing how the IP blocking works in snap.