Removing a few more left overs from /usr/share/snapwebsites/conf (since we support /etc/snapwebsites/snapwebsites.d/... it's not useful)
Changed the status shown of each backend service. If not selected as an active backend on that computer, it will be shown normally if disbaled.
Properly handle the case where a daemon status is "static" which we currently interpret as "disabled".
Fixed the snapbackend.postrm script, the purge if block was empty.