Removed the creation of the sessions directory under /var/lib/snapwebsites.
Properly delete the sessions data under /var/cache/www-snapwebsites.
Fixed the AJAX URI with a slash (i.e. "/snapmanager" instead of just "snapmanager".) So if the path includes a slash it still works.
Fixed the snapstop --service with quotes so it receives an empty parameter when $MAINPID is not defined and does not generate an error in that case.
Fixed a bug in installer.cpp, a section may already exist when we do not find the specified variable by the end of the file.