Added the manual_down field to the flag support so we can mark certain flags as requiring administrator intervention to get turned off.
Flag up two more cases in the content and editor plugins.
Moved the snapwatchdog_lib_SOURCE_DIR specification to the top CMakeLists.
Force link with snapwatchdog so it works as expected.
Correctly save the flag up in the antivirus code, also change priorty.
Added the --manual to the raise-flag tool.