Got a first version of the index plugin to work with the Document Data page.
Fixed the table name of the index plugin, it's "indexes" because "INDEX" is a reserved keyword.
Added a "start key" for the index because it will include many entries which ware not of interest to us when listing for a given object.
Removed the CRON action, at the moment the "index" plugin has no CRON anything. It's instantaneous.
Added a "DELETE" in the index_one_page() to make sure the reindex() function can correctly do its job (because the default function only adds new entries, it doesn't delete anything.)
Avoid the reindexing loop if the original key script is empty (i.e. there's not going to be any entries if there is no script in the first place!)
Changed the reindex() loop with a for() to make sure the removeAt(0) happens.
Added a link() function in our snap_expr so that way we can follow a link in our Snap! database.
Fixed a HUGE bug in content::modified_content_impl() -- I checked whether the server was ready and returned false if so, instead of test if it wasn't ready!
Small clean ups in QLockFile which we now use in our config so make sure two different processes can't do the save() simultaneously.
Various clean ups.