SNAP-705: more progress on the snapdatabase implementation.
Fixed the support for column identifiers, these are assigned after the entire tables and its extensions were loaded in memory.
Actually save the "block_size" so on load it's there.
Properly re-load array items (the set_array() doesn't work; the offset handling was wrong).
Properly set column identifiers when we have an existing schema (i.e. columns with the same name need to keep the exact same identifier).
Added propoer support for secondary indexes in the schemata.
Extended the secondary index definitions with a full "sort column" structure so we can have more than just a name (i.e. descending, not-null...)
Added a filter field to the secondary index because in most cases an index only includes a "few rows" not the entire database.
Avoid calling compile_script() when the script is the empty string and that's allowed/expected.
Properly transform secondary indexes to and from binary.
Fixed the complex types again (i.e. the context reads ALL of them from ALL the table files, then parses the tables).
Fixed various types, added missing definitions and updated the secondary-indexes in the tables.xsd definitions.
Enhanced the print out of the virtual buffer to include ASCII characters to the right.
Added a verification of the XML tables tests (xmllint).