SNAP-705: Started work to send row data. I fixed the cell, added some types, fixed the type verification with a function.
Optimized the set function of small integers.
Added support for a length to sorting columns.
Fixed "_oid" type, it had to use OID.
Added the "_schema_version" system column.
The columns & secondary indexes are process in sub-functions so it can be done in the table extensions as in the base table.
Fixed the tables.xsd so it works with our current tests. We run the xmllint against our test XML files.
Changed the list of filenames to use a set, that way it gets sorted so multiple runs have a greater chance to reproduce the same problem if one is present.