SNAP-719: Properly reply on errors.
The snaplistd would not properly send its reply to the process asking
for the info to be sent to the backend. (i.e. the snaplistd.cpp is
the one running on the backend accepting messages from the various
computers that may make changes to the database and require a list
update such as the front ends.)