In 21.04, the gjs is the linter.
I removed a couple of lines that were still looking for the linter the old
way and instead exclusively rely on the SnapJsLint.cmake implementation.

Also the new name had to be added to the list of names to search for and
I added the --jscomp_warning=lintChecks command line option.