Fixed a few more #include.
But I think that most of this code is lost... I will redesign and make use
of a REST accessible database instead of a backend. i.e. most of the work
will be done by clients instead (in the browser). The backend will just be
a "small" set of kickstart JS files which load everything else.